+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (unreleased)
 
 - Added `NN` module
+- Added `bleu_score` method
 
 ## 0.1.0 (2020-08-24)